Allow async runtime selection via redis crate feature flags, tests for tokio runtime, updated docs, version bump
