+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "redis-graph"
-version = "0.1.1"
+version = "0.2.0"
 authors = ["tompro <office@protom.eu>"]
 keywords = ["redis", "database", "graph"]
 description = "API for Redis graph database types."
@@ -13,9 +13,22 @@ edition = "2018"
 exclude = ["docker"]
 
 [dependencies]
-redis = "0.17.0"
+redis = { version = "0.19.0", optional = true }
+
+[features]
+default = ['redis']
+tokio-comp = ['redis/tokio-comp']
+async-std-comp = ['redis/async-std-comp']
 
 [dev-dependencies]
-tokio = { version = "0.2", features = ["full"] }
+tokio = { version = "1", features = ["rt"] }
 futures = "0.3.5"
 async-std = "1.6.0"
+
+[[test]]
+name = "test_async_std_commands"
+required-features = ['async-std-comp']
+
+[[test]]
+name = "test_async_tokio_commands"
+required-features = ['tokio-comp']
